+++ cuda/solver/gmres_mixed_kernels.cu
@@ -93,32 +93,33 @@ constexpr int default_update_dim = 32;
 
 // Specialization, so the Accessor can use the same function as regular pointers
 template <typename Type1, typename Type2>
-ReducedStorage3d<cuda_type<Type1>, cuda_type<Type2>> as_cuda_accessor(
-    ReducedStorage3d<Type1, Type2> acc)
+GKO_INLINE accessor::ReducedStorage3d<cuda_type<Type1>, cuda_type<Type2>>
+as_cuda_accessor(accessor::ReducedStorage3d<Type1, Type2> &acc)
 {
     return {as_cuda_type(acc.get_storage()), acc.get_stride0(),
             acc.get_stride1()};
 }
 
 template <typename Type1, typename Type2>
-ScaledReducedStorage3d<cuda_type<Type1>, cuda_type<Type2>> as_cuda_accessor(
-    ScaledReducedStorage3d<Type1, Type2> acc)
+GKO_INLINE accessor::ScaledReducedStorage3d<cuda_type<Type1>, cuda_type<Type2>>
+as_cuda_accessor(accessor::ScaledReducedStorage3d<Type1, Type2> &acc)
 {
     return {as_cuda_type(acc.get_storage()), acc.get_stride0(),
             acc.get_stride1(), as_cuda_type(acc.get_scale())};
 }
 
 template <typename Type1, typename Type2>
-ConstReducedStorage3d<cuda_type<Type1>, cuda_type<Type2>> as_cuda_accessor(
-    const ConstReducedStorage3d<Type1, Type2> &acc)
+GKO_INLINE accessor::ConstReducedStorage3d<cuda_type<Type1>, cuda_type<Type2>>
+as_cuda_accessor(const accessor::ConstReducedStorage3d<Type1, Type2> &acc)
 {
     return {as_cuda_type(acc.get_storage()), acc.get_stride0(),
             acc.get_stride1()};
 }
 
 template <typename Type1, typename Type2>
-ConstScaledReducedStorage3d<cuda_type<Type1>, cuda_type<Type2>>
-as_cuda_accessor(const ConstScaledReducedStorage3d<Type1, Type2> &acc)
+GKO_INLINE accessor::ConstScaledReducedStorage3d<cuda_type<Type1>,
+                                                 cuda_type<Type2>>
+as_cuda_accessor(const accessor::ConstScaledReducedStorage3d<Type1, Type2> &acc)
 {
     return {as_cuda_type(acc.get_storage()), acc.get_stride0(),
             acc.get_stride1(), as_cuda_type(acc.get_scale())};
@@ -280,7 +281,7 @@ void finish_arnoldi_CGS2(std::shared_ptr<const CudaExecutor> exec,
     const dim3 block_size_iters_single(singledot_block_size);
     size_type numReorth;
 
-    ReducedStorage3d<ValueType, ValueType> next_krylov_accessor{
+    accessor::ReducedStorage3d<ValueType, ValueType> next_krylov_accessor{
         next_krylov_basis->get_values(), stride_next_krylov,
         stride_next_krylov};
